requirements.txt
python-lambda
azure-functions
azure-storage
datetime
xlrd